http-api-source-ui/requirements.txt
quixstreams==3.23.2
flask
flask_cors
waitress
python-dotenv